+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed `--inline-threshold` deprecation warning in `embassy` examples.
 - Require safety documentation for unsafe blocks.
+- Fixed [FromRequest](https://docs.rs/picoserve/latest/picoserve/extract/trait.FromRequest.html) for `alloc::vec::Vec`, and by extension, `alloc::string::String`.
 
 ### Changes